Update test_portability and rename to test_distbuild
